feat : Assembly and Pluginification
Moved code to namespace, Folder and gave it an assembly, updated Friends assembly to include.
