+++ Assets/Plugins/SampleAuthentication.meta
