+++ Assets/Plugins/SampleAuthentication/Interfaces.meta
