+++ Assets/Plugins/SampleAuthentication/Samples.Authentication.asmdef
@@ -1,13 +1,9 @@
 {
-    "name": "FriendsSample",
+    "name": "Samples.Authentication",
     "rootNamespace": "",
     "references": [
-        "GUID:c3e847927e5747f9a371134e85820ab3",
-        "GUID:343deaaf83e0cee4ca978e7df0b80d21",
-        "GUID:6055be8ebefd69e48b49212b09b47b2f",
-        "GUID:2bafac87e7f4b9b418d9448d219b01ab",
-        "GUID:ce385344091ea4cfbb93b42ba4bb6c1e",
-        "GUID:fe25561d224ed4743af4c60938a59d0b"
+        "GUID:fe25561d224ed4743af4c60938a59d0b",
+        "GUID:5540e30183c82e84b954c033c388e06c"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
