+++ Assets/Plugins/SampleAuthentication/Samples.Authentication.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ce385344091ea4cfbb93b42ba4bb6c1e
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
