+++ Assets/Plugins/SampleAuthentication/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 354cbbf4855f04055905c2e95b8a0ec0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
