+++ Assets/Plugins/SampleAuthentication/Scripts/UGSPlayer.cs.meta
